Connection OK Debut de CompanyDAO

+++ src/main/java/com/excilys/computerdatabase/model/ICompany.java
@@ -2,11 +2,11 @@ package com.excilys.computerdatabase.model;
 
 public interface ICompany {
 	
-	public int getId();
+	public long getId();
 	
 	public String getName();
 	
-	public void setId(int id);
+	public void setId(long id);
 	
 	public void setName(String name);
